Created authorsFileTouches.py and scatterplot.py
michael_authorsFileTouches.py goes through the list generated by the
CollectFiles script and for each file in the list, extracts the authors
who contributed and the time at which they commited. Stores in a .csv
for scatterplot.py. scatterplot script cleans data and presents at as a
scatterplot of commits vs weeks.